backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.3.1
django-cors-headers==4.7.0
django-filter==25.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
urllib3==2.3.0
